@@ -4,12 +4,12 @@ version: 0.5.0
 homepage: https://github.com/yohom/demo_widgets
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=3.3.4 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  decorated_flutter: ^0.40.0-dev.1
+  decorated_flutter: ^0.49.0
   flutter_syntax_view: ^4.0.0
 
 dev_dependencies:
