@@ -40,7 +40,7 @@ class _ContinuousSettingState extends State<ContinuousSetting> {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: <Widget>[
-          Text(widget.head, style: Theme.of(context).textTheme.subtitle1),
+          Text(widget.head, style: Theme.of(context).textTheme.titleMedium),
           SPACE_8,
           Slider(
             value: _value,
@@ -79,7 +79,7 @@ class DiscreteSetting extends StatelessWidget {
       onSelected: onSelected,
       child: Padding(
         padding: const EdgeInsets.all(16),
-        child: Text(head, style: Theme.of(context).textTheme.subtitle1),
+        child: Text(head, style: Theme.of(context).textTheme.titleMedium),
       ),
       itemBuilder: (context) {
         return options
